Refactor #7 - extract remaining views, helpers, tidy up
- index and results are extracted
- perform_advanced_search and pagination become util helpers
- get s3 ingest log moves to aws helper
- other constants stored in views move to utils
- views are imported directly from their specific modules in urls
