+++ judgments/utils/paginator.py
@@ -1,0 +1,21 @@
+import math
+
+from caselawclient.Client import RESULTS_PER_PAGE
+
+def paginator(current_page, total):
+    size_per_page = RESULTS_PER_PAGE
+    number_of_pages = math.ceil(int(total) / size_per_page)
+    next_pages = list(
+        range(current_page + 1, min(current_page + 10, number_of_pages) + 1)
+    )
+
+    return {
+        "current_page": current_page,
+        "has_next_page": current_page < number_of_pages,
+        "next_page": current_page + 1,
+        "has_prev_page": current_page > 1,
+        "prev_page": current_page - 1,
+        "next_pages": next_pages,
+        "number_of_pages": number_of_pages,
+    }
+
